implement tests without fixture data
