+++ plugins/UserCountry/tests/System/ApiTest.php
@@ -9,7 +9,9 @@
 
 namespace Piwik\Plugins\UserCountry\tests\System;
 
+use DateTime;
 use Piwik\Plugins\UserCountry\tests\Fixtures\ManySitesManyVisitsWithGeoIp;
+use Piwik\API\Request;
 use Piwik\Tests\Framework\TestCase\SystemTestCase;
 
 /**
@@ -96,6 +98,73 @@ class ApiTest extends SystemTestCase
     public static function getPathToTestDirectory()
     {
         return dirname(__FILE__);
+    }
+
+    public function testGetCountry()
+    {
+        $dateTime = '2026-01-01 00:00:00';
+        $idSite = static::$fixture::createWebsite(
+            $dateTime,
+            $ecommerce = 1,
+        );
+
+        $this->createManyEcommerceOrders($idSite, $dateTime, 100);
+
+        $result = Request::processRequest('UserCountry.getCountry', [
+            'idSite' => $idSite,
+            'period' => 'year',
+            'date' => $dateTime,
+            'flat' => '1',
+        ]);
+
+        $resultWithSegment = Request::processRequest('UserCountry.getCountry', [
+            'idSite' => $idSite,
+            'period' => 'year',
+            'date' => $dateTime,
+            'flat' => '1',
+            'segment' => 'countryName==Italy',
+        ]);
+
+        foreach ($result->getRows() as $row) {
+            var_dump($row->getColumns());
+        }
+
+        foreach ($resultWithSegment->getRows() as $row) {
+            var_dump($row->getColumns());
+        }
+
+    }
+
+    public function createManyEcommerceOrders($siteId, $dateTime, $numberOfOrders)
+    {
+        static::$fixture::createSuperUser($removeExisting = true);
+
+        $visitorIds = [
+            'visit-1',
+            'visit-2',
+            'visit-3',
+        ];
+
+        $orderNumber = 1001;
+
+        $dateTimeObj = new DateTime($dateTime);
+        $interval = new \DateInterval('P1D');
+
+        for ($i = 0; $i < $numberOfOrders; $i++) {
+            $dateTimeStr = date_format($dateTimeObj, 'Y-m-d H:i:s');
+            $tracker = static::$fixture::getTracker($siteId, $dateTimeStr, $defaultInit = true, $useLocal = true);
+            // test both IP address and country code for country detection for segment
+            if ($i % 2 == 0) {
+                $tracker->setIp('151.100.101.92'); // Italy
+            } else {
+                $tracker->setCountry('it');
+            }
+            $tracker->setVisitorId(substr(md5($visitorIds[$i % 3]), $offset = 0, $tracker::LENGTH_VISITOR_ID));
+            $tracker->setTokenAuth(static::$fixture::getTokenAuth());
+            $tracker->addEcommerceItem('SKU-' . ($orderNumber + $i), 'Product ' . ($orderNumber + $i), 'Category', 100, 1);
+            static::$fixture::checkResponse($tracker->doTrackEcommerceOrder($orderNumber + $i, 111.11, 100, 11));
+            $dateTimeObj = date_add($dateTimeObj, $interval);
+        }
     }
 }
